# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  core.Currency:
    properties:
      id:
        type: integer
      iso_code:
        type: string
      minor_units:
        type: integer
      name:
        type: string
      symbol:
        type: integer
    type: object
  core.CurrencyCreateInput:
    properties:
      iso_code:
        type: string
      minor_units:
        type: integer
      name:
        type: string
      symbol:
        type: integer
    type: object
  core.CurrencyUpdateInput:
    properties:
      iso_code:
        type: string
      minor_units:
        type: integer
      name:
        type: string
      symbol:
        type: integer
    type: object
...
  core.User:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      middle_name:
        type: string
      password_hash:
        type: string
      phone_number:
        type: string
    type: object
  core.UserCreateInput:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      password_hash:
        type: string
      phone_number:
        type: string
    type: object
  core.UserUpdateInput:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
    type: object
...
info:
  contact: {}
  description: HTTP API for users, accounts, cards, credits, deposits, currencies,
    and exchange rates.
  title: Bank Backend API
  version: "1.0"
paths:
...
  /currencies:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.Currency'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List currencies
      tags:
      - currencies
    post:
      consumes:
      - application/json
      parameters:
      - description: Currency data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/core.CurrencyCreateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.Currency'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create currency
      tags:
      - currencies
  /currencies/{id}:
    delete:
      parameters:
      - description: Currency ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete currency
      tags:
      - currencies
    get:
      parameters:
      - description: Currency ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.Currency'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get currency by ID
      tags:
      - currencies
    put:
      consumes:
      - application/json
      parameters:
      - description: Currency ID
        in: path
        name: id
        required: true
        type: integer
      - description: Currency changes
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/core.CurrencyUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.Currency'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update currency
      tags:
      - currencies
  /currencies/iso/{iso_code}:
    get:
      parameters:
      - description: ISO code
        in: path
        name: iso_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.Currency'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get currency by ISO code
      tags:
      - currencies
  /currencies/symbol/{symbol}:
    get:
      parameters:
      - description: Currency symbol
        in: path
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.Currency'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get currency by symbol
      tags:
      - currencies
...
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/core.UserCreateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.User'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create user
      tags:
      - users
  /users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete user
      tags:
      - users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.User'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User changes
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/core.UserUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.User'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update user
      tags:
      - users
  /users/{id}/password:
    patch:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: New password
        in: formData
        name: hash_password
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Change user password
      tags:
      - users
  /users/email/{email}:
    get:
      parameters:
      - description: Email address
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.User'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get user by email
      tags:
      - users
  /users/phone/{phone_number}:
    get:
      parameters:
      - description: Phone number
        in: path
        name: phone_number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.User'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get user by phone number
      tags:
      - users
schemes:
- http
- https
swagger: "2.0"